Switched to SDL 2 and added Makefile
Due to a desire to access elementary shapes (lines, circles, rectangles,
etc) without creating images for them, I decided to convert my previous
work from SDL to SDL 2. I also added a makefile to help with program
compilation. As with the previous commits, much of the code so far
closely follows that of the Lazy Foo Productions SDL 2 Tutorial.